#390 removed all filters
