+++ elastic4s-core/src/main/scala/com/sksamuel/elastic4s/ElasticDsl.scala
@@ -3,7 +3,7 @@ package com.sksamuel.elastic4s
 import java.util.UUID
 
 import com.sksamuel.elastic4s.admin.{OpenIndexDefinition, TypesExistsDefinition, RefreshIndexDefinition, IndicesStatsDefinition, IndexExistsDefinition, GetSegmentsDefinition, GetTemplateDefinition, FlushIndexDefinition, DeleteIndexTemplateDefinition, FieldStatsDefinition, ClusterStatsDefinition, ClusterHealthDefinition, ClusterStateDefinition, ClusterSettingsDefinition, CloseIndexDefinition, ClearCacheDefinition, ClusterDsl, SnapshotDsl, IndexTemplateDsl, IndexAdminDsl, FieldStatsDsl}
-import com.sksamuel.elastic4s.anaylzers.{StemmerTokenFilter, SnowballTokenFilter, ShingleTokenFilter, TokenFilterDsl, TokenizerDsl, AnalyzerDsl}
+import com.sksamuel.elastic4s.anaylzers.{TokenFilterDsl, TokenizerDsl, AnalyzerDsl}
 import com.sksamuel.elastic4s.mappings.FieldType.{ObjectType, NestedType, TokenCountType, StringType, ShortType, LongType, IpType, IntegerType, GeoShapeType, DateType, DoubleType, GeoPointType, MultiFieldType, FloatType, CompletionType, BooleanType, ByteType, BinaryType, AttachmentType}
 import com.sksamuel.elastic4s.mappings.{ObjectFieldDefinition, NestedFieldDefinition, TimestampDefinition, DynamicTemplateDefinition, StringFieldDefinition, PutMappingDefinition, MappingDefinition, GetMappingDefinition, DeleteMappingDefinition, TypeableFields, FieldDefinition, MappingDsl}
 
@@ -546,9 +546,3 @@ trait ElasticDsl
 }
 
 object ElasticDsl extends ElasticDsl
-
-//mapping char filter
-//
-//htmlStrip char filter
-//
-//patternReplace char filter
